chore: move MobileInputMode to utilities/mobile
