+++ _examples/widget_demos/cut_copy_paste/main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/ebitenui/ebitenui"
 	"github.com/ebitenui/ebitenui/image"
 	"github.com/ebitenui/ebitenui/input"
-	"github.com/ebitenui/ebitenui/internal/jsUtil"
+	"github.com/ebitenui/ebitenui/utilities/mobile"
 	"github.com/ebitenui/ebitenui/widget"
 	"github.com/hajimehoshi/ebiten/v2"
 	"github.com/hajimehoshi/ebiten/v2/inpututil"
@@ -59,7 +59,7 @@ func main() {
 		),
 
 		// Set the keyboard type when opened on mobile devices.
-		widget.TextInputOpts.MobileInputMode(jsUtil.TEXT),
+		widget.TextInputOpts.MobileInputMode(mobile.TEXT),
 
 		// Set the Idle and Disabled background image for the text input.
 		// If the NineSlice image has a minimum size, the widget will use that or
