+++ utilities/mobile/inputmode.go
@@ -1,0 +1,13 @@
+package mobile
+
+type MobileInputMode string
+
+const (
+	TEXT      = MobileInputMode("text")
+	DECIMAL   = MobileInputMode("decimal")
+	NUMERIC   = MobileInputMode("numeric")
+	TELEPHONE = MobileInputMode("tel")
+	SEARCH    = MobileInputMode("search")
+	EMAIL     = MobileInputMode("email")
+	URL       = MobileInputMode("url")
+)
